+++ sprofiler/src/lib.rs
@@ -1,3 +1,8 @@
-mod bpf;
+pub mod bpf;
+pub mod command;
+pub mod dynamic;
+pub mod ioutil;
+pub mod ociutil;
+pub mod profile_util;
+pub mod r#static;
 pub mod syscalls;
-pub use bpf::*;
